Amazon Product Recommender
This project entailed using Python to scrape 4,286 reviews of 146 headphones where the name of the headphones, the review, and the rating were collected. The review was then used to determine the top three most mentioned attributes. Using cosine similarity, reviews, and spacy similarity, an overall rating was determined to help recommend headphones that most catered to specific attributes.